First step to incremental database

+++ import_database.php
@@ -1,47 +1,55 @@
 <?php
 require_once 'boot.php';
 
-mysql_query('DROP TABLE node');
-mysql_query('DROP TABLE node_halt');
-//mysql_query('DROP TABLE node_tags');
-mysql_query('CREATE TABLE node ( id BIGINT(20) unsigned not null PRIMARY KEY, lat FLOAT(9,7), lon FLOAT(9,7) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci' );
-//mysql_query('CREATE TABLE node_tags ( node_id BIGINT(20) not null, field VARCHAR(255), value VARCHAR(255), UNIQUE KEY node_key (node_id, field), INDEX field (field), INDEX node_id (node_id) )');
-mysql_query('CREATE TABLE suburb ( node_id BIGINT(20) PRIMARY KEY, name VARCHAR(255), is_in VARCHAR(255), INDEX name (name), INDEX is_in (is_in) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci');
+$keep_data = FALSE;
+$suburbs = array();
+if ($keep_data === FALSE) {
+	mysql_query('DROP TABLE node');
+	mysql_query('DROP TABLE suburb');
+	//mysql_query('DROP TABLE node_tags');
+	mysql_query('CREATE TABLE node ( id BIGINT(20) unsigned not null PRIMARY KEY, lat FLOAT(9,7), lon FLOAT(9,7) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci' ) or die(mysql_error());
+	//mysql_query('CREATE TABLE node_tags ( node_id BIGINT(20) not null, field VARCHAR(255), value VARCHAR(255), UNIQUE KEY node_key (node_id, field), INDEX field (field), INDEX node_id (node_id) )');
+	mysql_query('CREATE TABLE suburb ( node_id BIGINT(20) PRIMARY KEY, name VARCHAR(255), is_in VARCHAR(255), INDEX name (name), INDEX is_in (is_in) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci') or die(mysql_error());
 
-mysql_query('DROP TABLE way');
-mysql_query('DROP TABLE way_nodes');
-mysql_query('DROP TABLE railway');
-mysql_query('DROP TABLE railway_halts');
-//mysql_query('DROP TABLE way_tags');
-mysql_query('CREATE TABLE way ( id BIGINT(20) unsigned not null PRIMARY KEY, name VARCHAR(255), street_id BIGINT(20), INDEX name (name), INDEX street_id (street_id) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci');
-mysql_query('CREATE TABLE way_nodes ( way_id BIGINT(20) UNSIGNED NOT NULL, node_id BIGINT(20) UNSIGNED NOT NULL, INDEX way_id (way_id), INDEX node_id (node_id) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci');
-mysql_query('CREATE TABLE railway ( id BIGINT(20) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY, way_id BIGINT(20) UNSIGNED NOT NULL, operator VARCHAR(255), UNIQUE KEY way_id (way_id) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci');
-mysql_query('CREATE TABLE railway_halts ( node_id BIGINT(20) PRIMARY KEY, name VARCHAR(255), INDEX node_id (node_id) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci');
-//mysql_query('CREATE TABLE way_tags ( way_id BIGINT(20) UNSIGNED NOT NULL, field VARCHAR(255), value VARCHAR(255), UNIQUE KEY way_key (way_id, field), INDEX field (field), INDEX way_id (way_id)  )');
+	mysql_query('DROP TABLE way');
+	mysql_query('DROP TABLE way_nodes');
+	mysql_query('DROP TABLE railway');
+	mysql_query('DROP TABLE railway_halts');
+	//mysql_query('DROP TABLE way_tags');
+	mysql_query('CREATE TABLE way ( id BIGINT(20) unsigned not null PRIMARY KEY, name VARCHAR(255), street_id BIGINT(20), INDEX name (name), INDEX street_id (street_id) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci') or die(mysql_error());
+	mysql_query('CREATE TABLE way_nodes ( way_id BIGINT(20) UNSIGNED NOT NULL, node_id BIGINT(20) UNSIGNED NOT NULL, INDEX way_id (way_id), INDEX node_id (node_id) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci') or die(mysql_error());
+	mysql_query('CREATE TABLE railway ( id BIGINT(20) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY, way_id BIGINT(20) UNSIGNED NOT NULL, operator VARCHAR(255), UNIQUE KEY way_id (way_id) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci') or die(mysql_error());
+	mysql_query('CREATE TABLE railway_halts ( node_id BIGINT(20) PRIMARY KEY, name VARCHAR(255), INDEX node_id (node_id) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci') or die(mysql_error());
+	//mysql_query('CREATE TABLE way_tags ( way_id BIGINT(20) UNSIGNED NOT NULL, field VARCHAR(255), value VARCHAR(255), UNIQUE KEY way_key (way_id, field), INDEX field (field), INDEX way_id (way_id)  )');
 
-mysql_query('DROP TABLE train');
-mysql_query('DROP TABLE train_stations');
-mysql_query('DROP TABLE bus');
-mysql_query('DROP TABLE relation_nodes');
-mysql_query('DROP TABLE relation_ways');
-//mysql_query('DROP TABLE relation_tags');
-mysql_query('CREATE TABLE train ( id BIGINT(20) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY, way_id BIGINT(20) UNSIGNED NOT NULL, operator VARCHAR(255), UNIQUE KEY way_id (way_id) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci');
-mysql_query('CREATE TABLE train_stations ( node_id BIGINT(20) PRIMARY KEY, name VARCHAR(255), INDEX node_id (node_id) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci');
-mysql_query('CREATE TABLE bus ( relation_id BIGINT(20) UNSIGNED NOT NULL PRIMARY KEY, name VARCHAR(255), ref VARCHAR(255), operator VARCHAR(255), INDEX relation_id (relation_id) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci'); 
-mysql_query('CREATE TABLE relation_nodes ( relation_id BIGINT(20) UNSIGNED NOT NULL, node_id BIGINT(20) UNSIGNED NOT NULL, INDEX relation_id (relation_id), INDEX node_id (node_id) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci');
-mysql_query('CREATE TABLE relation_ways ( relation_id BIGINT(20) UNSIGNED NOT NULL, way_id BIGINT(20) UNSIGNED NOT NULL, ordering INT(1) UNSIGNED, INDEX relation_id (relation_id), INDEX way_id (way_id) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci');
-//mysql_query('CREATE TABLE relation_tags ( relation_id BIGINT(20) UNSIGNED NOT NULL, field VARCHAR(255), value VARCHAR(255), UNIQUE KEY relation_key (relation_id, field, INDEX field (field), INDEX relation_id (relation_id)  )');
+	mysql_query('DROP TABLE train');
+	mysql_query('DROP TABLE train_stations');
+	mysql_query('DROP TABLE bus');
+	mysql_query('DROP TABLE relation_nodes');
+	mysql_query('DROP TABLE relation_ways');
+	//mysql_query('DROP TABLE relation_tags');
+	mysql_query('CREATE TABLE train ( id BIGINT(20) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY, way_id BIGINT(20) UNSIGNED NOT NULL, operator VARCHAR(255), UNIQUE KEY way_id (way_id) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci') or die(mysql_error());
+	mysql_query('CREATE TABLE train_stations ( node_id BIGINT(20) PRIMARY KEY, name VARCHAR(255), INDEX node_id (node_id) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci') or die(mysql_error());
+	mysql_query('CREATE TABLE bus ( relation_id BIGINT(20) UNSIGNED NOT NULL PRIMARY KEY, name VARCHAR(255), ref VARCHAR(255), operator VARCHAR(255), INDEX relation_id (relation_id) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci') or die(mysql_error()); 
+	mysql_query('CREATE TABLE relation_nodes ( relation_id BIGINT(20) UNSIGNED NOT NULL, node_id BIGINT(20) UNSIGNED NOT NULL, INDEX relation_id (relation_id), INDEX node_id (node_id) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci') or die(mysql_error());
+	mysql_query('CREATE TABLE relation_ways ( relation_id BIGINT(20) UNSIGNED NOT NULL, way_id BIGINT(20) UNSIGNED NOT NULL, ordering INT(1) UNSIGNED, INDEX relation_id (relation_id), INDEX way_id (way_id) ) DEFAULT CHARACTER SET utf8 COLLATE utf8_general_ci') or die(mysql_error());
+	//mysql_query('CREATE TABLE relation_tags ( relation_id BIGINT(20) UNSIGNED NOT NULL, field VARCHAR(255), value VARCHAR(255), UNIQUE KEY relation_key (relation_id, field, INDEX field (field), INDEX relation_id (relation_id)  )');
 
-mysql_query('DROP TABLE street');
-mysql_query('DROP TABLE street_suburbs');
-mysql_query('CREATE TABLE street ( id BIGINT(20) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY, name VARCHAR(255), full_name VARCHAR(255) )'); // street is an entity we create, so it must be autoincrement unlike ways, nodes and relations
-mysql_query('CREATE TABLE street_suburbs ( street_id BIGINT(20), suburb_id BIGINT(20), INDEX street_id (street_id), INDEX suburb_id (suburb_id) )');
+	mysql_query('DROP TABLE street');
+	mysql_query('DROP TABLE street_suburbs');
+	mysql_query('CREATE TABLE street ( id BIGINT(20) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY, name VARCHAR(255), full_name VARCHAR(255) )'); // street is an entity we create, so it must be autoincrement unlike ways, nodes and relations
+	mysql_query('CREATE TABLE street_suburbs ( street_id BIGINT(20), suburb_id BIGINT(20), INDEX street_id (street_id), INDEX suburb_id (suburb_id) )');
+} else {
+	$suburbs_query = mysql_query('SELECT node.id, node.lat, node.lon FROM suburb JOIN node ON suburb.node_id = node.id');
+	while ($suburb = mysql_fetch_assoc($suburbs_query)) {
+		$suburbs[$suburb['id']] = array($suburb['lat'], $suburb['lon']);
+	}
+}
 
 //$data = file_get_contents('sample');
 //$simplexml = simplexml_load_string($data);
 $simplexml = simplexml_load_file('sample');
 
-$suburbs = array();
 foreach ($simplexml as $node) {
 	if ($node->getName() == 'node') {
 		$node_id = $lat = $lon = NULL;
@@ -53,6 +61,11 @@ foreach ($simplexml as $node) {
 //			else $extradata[$key] = $value;
 		}
 		if ($node_id !== NULL && $lat !== NULL && $lon !== NULL) {
+			if (isset($suburbs[$node_id])) unset($suburbs[$node_id]);
+			mysql_query('DELETE FROM node WHERE id = ' . $node_id);
+			mysql_query('DELETE FROM suburb WHERE node_id = ' . $node_id);
+			mysql_query('DELETE FROM train_stations WHERE node_id = ' . $node_id);
+			mysql_query('DELETE FROM railway_halts WHERE node_id = ' . $node_id);
 			mysql_query('INSERT INTO node (id, lat, lon) VALUES (' . $node_id . ', '. $lat . ',' . $lon . ')');
 //			foreach ($extradata as $key => $value) {
 //				mysql_query('INSERT INTO node_tags (node_id, field, value) VALUES (' . $node_id . ', "' . mysql_real_escape_string($key) . '", "' . mysql_real_escape_string($value) . '")');
@@ -121,6 +134,10 @@ foreach ($simplexml as $node) {
 				if ($is_name) $name = $v;
 			}
 			if (($is_highway || $railway !== NULL || $train !== NULL) && $name !== NULL && count($nodes) > 0) {
+				mysql_query('DELETE FROM way WHERE id = ' . $way_id);
+				mysql_query('DELETE FROM railway WHERE way_id = ' . $way_id);
+				mysql_query('DELETE FROM train WHERE way_id = ' . $way_id);
+				mysql_query('DELETE FROM way_nodes WHERE way_id = ' . $way_id);
 				mysql_query('INSERT INTO way (id, name) VALUES (' . $way_id . ', "' . mysql_real_escape_string($name) . '")');
 				if ($railway !== NULL) mysql_query('INSERT INTO railway (way_id, operator) VALUES (' . $way_id . ', "' . $operator . '")');
 				if ($train !== NULL) mysql_query('INSERT INTO train (way_id, operator) VALUES (' . $way_id . ', "' . $operator . '")');
@@ -154,6 +171,9 @@ foreach ($simplexml as $node) {
 				}
 			}
 			if ($route == 'bus') {
+				mysql_query('DELETE FROM bus WHERE relation_id = ' . $way_id);
+				mysql_query('DELETE FROM relation_ways WHERE relation_id = ' . $way_id);
+				mysql_query('DELETE FROM relation_nodes WHERE relation_id = ' . $way_id);
 				mysql_query('INSERT INTO bus (relation_id, name, operator, ref) VALUES (' . $relation_id . ', "' . mysql_real_escape_string($name) . '", "' . mysql_real_escape_string($operator) . '", "' . mysql_real_escape_string($ref) . '")');
 				$i = 0;
 				foreach ($ways as $way)
@@ -165,6 +185,9 @@ foreach ($simplexml as $node) {
 	}
 }
 
+mysql_query('TRUNCATE street');
+mysql_query('TRUNCATE street_suburbs');
+mysql_query('UPDATE way SET street_id = NULL');
 while (true) {
 	$ways = mysql_query('SELECT way.id, way.name FROM way LEFT JOIN railway ON way.id = railway.way_id WHERE railway.id IS NULL AND way.street_id IS NULL ORDER BY way.name ASC LIMIT 500');
 	$rows = mysql_num_rows($ways);
